CORE-1484 Extract logic in InstantLaunchButton into separate component
The navigation drawer when displaying "first class" instant launches will need this same launch service call and will also need to display the Submission dialog to the user.  This way any clickable element can reuse this same logic.
